PAE-1790: stop shared code importing feature internals
Non-production database seeding moves to the composition root under
src/server/seed/, so the Mongo plugin no longer builds the organisations
repository itself and takes the seeding step as an option instead.

Summary-log metrics move to the summary-logs application layer, and the
organisation-scoped Defra ID role resolution moves to the organisations
application layer, injected into the JWT strategy by the auth plugin.
