+++ src/common/helpers/plugins/mongo-db-plugin.js
@@ -1,6 +1,4 @@
 import { createMongoClient } from '#common/helpers/mongo-client.js'
-import { createOrganisationsRepository } from '#repositories/organisations/mongodb.js'
-import { isProductionEnvironment } from '#root/config.js'
 import { LockManager } from 'mongo-locks'
 import {
   LOGGING_EVENT_ACTIONS,
@@ -8,8 +6,7 @@ import {
 } from '../../enums/index.js'
 import {
   createFormCollections,
-  createLockManagerIndex,
-  createSeedData
+  createLockManagerIndex
 } from '../collections/create-update.js'
 
 export const mongoDbPlugin = {
@@ -18,6 +15,12 @@ export const mongoDbPlugin = {
     version: '1.0.0',
     /**
      * @param {import('../../hapi-types.js').HapiServer} server
+     * @param {{
+     *   mongoUrl: string,
+     *   mongoOptions: object,
+     *   databaseName: string,
+     *   seedDatabase: (db: import('mongodb').Db) => Promise<void>
+     * }} options
      */
     register: async function (server, options) {
       server.logger.info({
@@ -39,15 +42,7 @@ export const mongoDbPlugin = {
 
       await createFormCollections(db)
       await createLockManagerIndex(db)
-
-      const organisationsRepositoryFactory =
-        await createOrganisationsRepository(db)
-
-      await createSeedData(
-        db,
-        isProductionEnvironment,
-        organisationsRepositoryFactory()
-      )
+      await options.seedDatabase(db)
 
       server.logger.info({
         message: `MongoDb connected to ${options.databaseName}`,
